requirements.txt
Django==1.5
distribute==0.6.34
dj-database-url==0.2.1
psycopg2==2.4.6
wsgiref==0.1.2
django-ajax-selects==1.2.5